+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add the ability to always remember the selected environment at the project level
 - Added a new settings modal for managing this new project-level setting, along with the app-wide default for all projects.
 - Remove prepended "Eyas" title from the app window title
+- Do not display data: URLs in the app window title
 
 # v26.2.34056