+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v26.x
+# v26.3.19614
 
 - Global config variables are now prepended with an underscore
 - Add support for getting environment data from the `_env` variable